backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-cloud-firestore
python-multipart
httpx